refactor(PurchaseLottoForm): play 메서드 form 컴포넌트로 이동
